feat: sync also odometry (BROKEN)

+++ build_and_run.sh
@@ -1,3 +1,7 @@
 # This is for making clangd finding all the includes, so that LSP find no errors for missing ros libraries
 catkin_make -DCMAKE_EXPORT_COMPILE_COMMANDS=1
-ln -s build/compile_commands.json compile_commands.json
+
+if [ ! -f compile_commands.json ]; then
+    ln -s build/compile_commands.json compile_commands.json
+fi
+roslaunch launch/robotics_project.launch
